move LinksEntry to entry_links.go
