+++ entry.go
@@ -21,7 +21,6 @@ const (
 
 	// Entry type formats
 
-	DataFormatLinks   = "links"
 	DataFormatJSON    = "json"
 	DataFormatString  = "string"
 	DataFormatRawJS   = "js"
@@ -33,19 +32,6 @@ const (
 	Partial = "partial"
 	Private = "private"
 )
-
-// LinksEntry holds one or more links
-type LinksEntry struct {
-	Links []Link
-}
-
-// Link structure for holding meta tagging of linking entry
-type Link struct {
-	LinkAction string // StatusAction (either AddAction or DelAction)
-	Base       string // hash of entry (perhaps elsewhere) to which we are attaching the link
-	Link       string // hash of entry being linked to
-	Tag        string // tag
-}
 
 // EntryDef struct holds an entry definition
 type EntryDef struct {
@@ -200,17 +186,5 @@ func (d *EntryDef) BuildJSONSchemaValidatorFromString(schema string) (err error)
 	}
 	validator.v.SetName(d.Name)
 	d.validator = validator
-	return
-}
-
-func (ae *LinksEntry) ToJSON() (encodedEntry string, err error) {
-	var j []byte
-	j, err = json.Marshal(ae)
-	encodedEntry = string(j)
-	return
-}
-
-func LinksEntryFromJSON(j string) (entry LinksEntry, err error) {
-	err = json.Unmarshal([]byte(j), &entry)
 	return
 }
